OcTreeBaseImpl: add setNodeChild interface
There are cases where trees need to be able to alter child pointers.
Provide an interface to do so.

One such case is when replacing a node in a recursion call and
returning a new node pointer. In this case, the caller needs to
modify the children pointer for the child.