OccupancyOcTreeBase: add setTreeValues()
Add a method to set many values at once from another tree.

Add a bounds-limited version of this method as well to make updating
portions of a tree from another tree possible.

Provide a method argument, maximum_only, to limit the set values to
only the maximum of the value stored in the tree being update and the
values in the other tree.

Provide a method argument, delete_first, to delete any regions to be
updated from this tree first. This argument is useful when using a
bounds-limiting tree to update a tree from the parts of another tree
that have changed.