test_color_tree: avoid adding points near boundary
Previously, the calculation for the free cells was setting a value
very near the resolution. Depending on compiler options, it is
possible to get this test to fail because the calculations end up
rounding on the other side of the boundary and creating more cells
than intended. Fix this by aligning the free cells creation to the
resoultion but off by a centimeter, similar to the occupied cells
insertion step.