deleteAABB: delete an axis-aligned bounding box
Add methods to delete cells either in or out of an axis-aligned
bounding box. This is a common operation for users of the octomap
library, and accelerating it by not visiting nodes to be deleted is
very useful in certain applications, like bounds-limited rolling
costmaps.

Also adds the ability to register a delection callback during node
deletion. This callback is called for every leaf node that is deleted.