node change callback
Add a way to register a callback to be called when a node is created
or when a node's value changes.

Add updateNodeLogOddsAndTrackChanges which can be called to update
the specified node's log odds in such a way that a change will
be tracked. This can be useful as a part of a copy function in
setTreeValues, for instance, where changes are not normally tracked.