Use a boost pool to allocate nodes and children
Use a boost pool allocated to each tree instance to allocated
nodes and children pointer arrays. This vastly speeds up time
spent allocating and freeing nodes, as a pool is more efficient
for small allocations of fixed size.

As a part of this change, delete the copy constructor for nodes.
It is no longer possible for a node to copy itself, as it does
not have access to the correct tree allocator. Because the tree
already must manage the children pointer, the copy constructor
was a relic anyway. This means any third party code must remove
any copy constructors they have defined for their nodes.