Use a boost pool to allocate nodes and children
Use a boost pool allocated in each tree instance to allocate nodes
and children pointer arrays. This vastly speeds up time spent
allocating and freeing nodes, as a pool is more efficient for small
allocations of fixed size.

As a part of this change, delete the copy constructor for nodes. It
is no longer possible for a node to copy itself, as it does not have
access to the correct tree allocator. Because the tree already must
manage the children pointer, the copy constructor was a relic anyway.
This means any third party code must remove any copy constructors
they have defined for their nodes.

For new trees, set the initial pool allocation to 128 to get a larger
initial chunk and reduce the number of calls needed to the allocator.
For copied trees, set the initial pool size to the tree size we are
copying so only one initial allocation must be done.