nodeHasChildren: only check children pointer
Make OcTreeBaseImpl always detect when a node has lost its last child.
When a node loses its children, delete the children pointer and
set it to NULL. Make nodeHasChildren simply check the children pointer
only to know if a node has children. This results in significant speed
up to tree walking at the small cost of an extra step when deleting.
Most workloads walk the tree much more than deleting nodes, so this is
a net improvement.