deleteNode: be willing to expand root
Due to the previous bug, another bug was present in the code.
In deleteNode, the root would not be expanded. However, this
is incorrect, it is possible to have a tree which represents
the whole universe be pruned to a single node at the root.
In such cases, expand the node just as any other inner node.