change key_type to uint32_t, support trees of varying depth
Make the write routines write the tree depth as a tree parameter.

Make the read routines backwards compatable, able to read old trees
without a depth specified by using the old default depth of 16.

Fix the many places the key size was fixed either explicitly or
implicitly.

Fix some places that integer arthemtic now overflows.

Fix any place when a bit shift greater than or equal to the bit width
of the key_type is performed. Such a shift is undefined. Instead,
explicitly code the value used in such a circumstance as zero. These
kinds of shifts happen when walking the tree. Add the
computeCenterOffsetKey function to centralize most of the computation
using bit shifts.

Fix test cases to not assume anything about the key size or tree depth.

Add setTreeDepth() method which alters the tree structure (if present)
to a tree of the new depth. Data within the bounds of the new tree
depth is preserved by re-rooting the tree appropriately.

Add a non-const version of setTreeValues() which alters the inputs to
have depths which match the called-on tree. setTreeValues only works
correctly when all the tress have the same depth and indexing keys.
Have the const version of setTreeValues() error-out if the trees
differ in depth or resolution.

Fix various places in octovis with hard-coded depth.
Change the UI element maximums depending on the bit-width of keys.

Have adjustKeyAtDepth return the input key if the passed in depth is
larger than the tree depth.

Have search clamp the search depth to the tree depth.