world coordinates float => double
Turn all world coordinates from float into double.
Float precision is not good enough with 32-bit keys when operating
nearer the edge of the map.
Double precision is actually faster on most modern processors.

Fix broken test cases by this change, due to differences in float rounding
giving different OcTreeKey's. Move such coordinates away from the
resolution boundary.

Also, fix the ray test to not project a ray from the center of the world
to the edge, which on a 32-bit map is a long way off. Now that there
is double precision, adjust the coordinates from the edge of the world
slightly towards the center and test casting the ray into the boundary
from near the boundary, resulting in a test with reasonable runtime.

Finally, fix the test macros to keep order of operations by adding
parenthesis around the macro arguments.