fix off-by-one error in bbx iterator with max depth
If the tree is KEY_BIT_WIDTH deep, and the bit width is 32 bits, the
size calculation in the octree bbx iterator is off by one, resulting
in caluculating the wrong max key (since the value will not fit in
the underlying integer). Fix this by adding one less to the size
going up. This also keeps us from unnecessarily descending parts of
the tree that were really out of the bounding box (but just on the
maximum edge).