fix node value callback memory leak
When a node value callback was set, change detection had to be
enabled. Users of the class may not realize this implication causing
them to leak memory by not freeing the resetting the changed nodes
set, slowly accumulating a huge number of change nodes on large maps.

Fix this by not forcing change detection when a node value change
callback is used. Instead, if a node value change callback is set,
call it, but do not track changed nodes unless change detection is
also set.