setTreeValues: fix bug with custom copy function
The intention of the custom copy function was to also be able to skip
copying if the node was newly created, to implement an invalid-only copy
(so values from the value tree would never override any existing values
in the current tree, by only copying values when node_just_created was
true). This, however, did not work when the value tree contained pruned
nodes, as the code assumed that the value tree would replace the current
tree if maximum_only was not set. Fix this by also descending the
current tree if a custom copy function is set so it can decide whether
or not to copy values over from the value tree to the current tree.

Add a test case to cover this use case of the custom copy function to
ensure it works properly.