search: add optional found_depth return by pointer
Add an optional found_depth return by pointer. If the found_depth
pointer is set, the depth the search stopped at will be set to what
found_depth points to. This way the caller can determine if the search
ended at the requested depth if that is important to the caller.